+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `{Global,}Tlsf::new`
 - `FlexTlsf::new` as a `const fn`
 - `ConstDefault` implementation for `Tlsf`
+- `Tlsf::iter_blocks` (unstable), which lets you iterate through memory blocks for diagnostic purposes.
 
 ### Removed